aura_aegis_sim/requirements.txt
streamlit>=1.32.0
plotly>=5.19.0
numpy>=1.26.0
pandas>=2.2.0
torch>=2.2.0
scikit-learn>=1.4.0
onnxruntime>=1.17.0
cryptography>=42.0.0
joblib>=1.3.0
